requirements.txt
chromadb==1.5.7
docx2txt==0.9
groq==0.37.1
langchain==1.2.15
langchain-chroma==1.1.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.31
langchain-groq==1.1.2
langchain-huggingface==1.2.1
langchain-openai==1.1.14
langchain-text-splitters==1.1.2
PyMuPDF==1.27.2.2
python-dotenv==1.2.2
sentence-transformers==5.4.1
streamlit==1.56.0
torchvision==0.26.0
